+++ propagation/src/propagation/ssw_2d.py
@@ -56,7 +56,7 @@ def ssw_2d(u_0, config, n_refraction, ii_vect_relief):
 
     # --- Creation of the apodisation window --- # @todo Code other apodisation windows
     # along z
-    n_apo_z = np.int64(config.apo_z * config.N_z)                                                                     #change
+    n_apo_z = np.int32(config.apo_z * config.N_z)                                                                     #change
     apo_window_z = apodisation_window(config.apo_window, n_apo_z)
     # ------------------------------------------ #
 
@@ -90,7 +90,7 @@ def ssw_2d(u_0, config, n_refraction, ii_vect_relief):
 
     # --- Sizes of the apodisation and image layers --- #
     if config.ground == 'PEC' or config.ground == 'Dielectric':
-        n_im = np.int64(np.round(config.N_z * config.image_layer))                                                    #change
+        n_im = np.int32(np.round(config.N_z * config.image_layer))                                                    #change
         remain_im = n_im % 2**config.wv_L
         if remain_im != 0:
             n_im += 2**config.wv_L - remain_im
