+++ propagation/src/propagation/wwp_2d.py
@@ -89,7 +89,7 @@ def wwp_2d(u_0, config, n_refraction, ii_vect_relief):
     np.save('./outputs/dictionary', dictionary)
     # --- Sizes of the apodisation and image layers --- #
     if config.ground == 'PEC' or config.ground == 'Dielectric':
-        n_im = np.int64(np.round(config.N_z * config.image_layer))                                                    #change
+        n_im = np.int32(np.round(config.N_z * config.image_layer))                                                    #change
         remain_im = n_im % 2**config.wv_L
         if remain_im != 0:
             n_im += 2**config.wv_L - remain_im
@@ -98,7 +98,7 @@ def wwp_2d(u_0, config, n_refraction, ii_vect_relief):
         n_im = 0
     config.N_im = n_im
 
-    n_apo_z = np.int64(config.apo_z * config.N_z)                                                                     #change
+    n_apo_z = np.int32(config.apo_z * config.N_z)                                                                     #change
     remain_apo = n_apo_z % 2 ** config.wv_L
     if remain_apo != 0:
         n_apo_z += 2 ** config.wv_L - remain_apo
